fix: rename file Props to interface, added in main index all components
