+++ src/components/CircleIconBtn/index.ts
@@ -1,5 +1,5 @@
 import { MetCircleIconBtn } from "./CircleIconBtn";
-import { MetCircleIconBtnProps } from "./CircleIconBtnProps";
+import { MetCircleIconBtnProps } from "./CircleIconBtn.interface";
 import { SizeCircleIconBtn } from "./sizeCircleIconBtn.enum";
 
 export { MetCircleIconBtnProps, SizeCircleIconBtn };
