+++ src/components/Icon/index.ts
@@ -1,5 +1,5 @@
 import { MetIcon } from "./Icon";
-import { MetIconProps } from "./IconProps";
+import { MetIconProps } from "./Icon.interface";
 import { icons } from "./icons";
 
 export { MetIconProps, icons };
