+++ src/components/PromptInfo/PromptInfo.interface.ts
