+++ src/components/PromptInfo/index.ts
@@ -1,5 +1,5 @@
 import { MetPromptInfo } from "./PromptInfo";
-import { MetPromptInfoProps } from "./PromptInfoProps";
+import { MetPromptInfoProps } from "./PromptInfo.interface";
 import { SizeCircleIconBtn } from "../CircleIconBtn/sizeCircleIconBtn.enum";
 
 export { MetPromptInfoProps, SizeCircleIconBtn };
