+++ src/components/RectangleIconBtn/index.ts
@@ -1,5 +1,5 @@
 import { MetRectangleIconBtn } from "./RectangleIconBtn";
-import { MetRectangleIconBtnProps } from "./RectangleIconBtnProps";
+import { MetRectangleIconBtnProps } from "./RectangleIconBtn.interface";
 import { TypesRectangleIconBtn } from "./typesRectangleIconBtn.enum";
 
 export { MetRectangleIconBtnProps, TypesRectangleIconBtn };
