+++ src/components/RectangleTextBtn/RectangleTextBtn.interface.ts
