+++ src/components/SquareIconCheckbox/SquareIconCheckbox.interface.ts
