+++ src/components/Tag/TagButton/index.ts
@@ -1,5 +1,5 @@
 import { MetTagButton } from "./TagButton";
-import { MetTagButtonProps } from "./TagButtonProps";
+import { MetTagButtonProps } from "./TagButton.interface";
 
 export { MetTagButtonProps };
