+++ src/components/Toggle/Toggle.interface.ts
