+++ src/components/index.ts
@@ -1,3 +1,6 @@
+export type { MetCheckboxProps } from "./Checkbox";
+export { default as MetCheckbox } from "./Checkbox";
+
 export type { MetCircleBtnProps } from "./CircleBtn";
 export { default as MetCircleBtn } from "./CircleBtn";
 
@@ -5,14 +8,26 @@ export { SizeCircleIconBtn } from "./CircleIconBtn";
 export type { MetCircleIconBtnProps } from "./CircleIconBtn";
 export { default as MetCircleIconBtn } from "./CircleIconBtn";
 
-export type { MetPromptInfoProps } from "./PromptInfo";
-export { default as MetPromptInfo } from "./PromptInfo";
+export { TypesDatePicker } from "./DatePicker";
+export type { MetDatePickerProps } from "./DatePicker";
+export { default as MetDatePicker } from "./DatePicker";
 
 export type { MetIconProps } from "./Icon";
 export { default as MetIcon } from "./Icon";
 
+export type { MetPromptInfoProps } from "./PromptInfo";
+export { default as MetPromptInfo } from "./PromptInfo";
+
+export type { MetRadioBtnProps } from "./RadioBtn";
+export { default as MetRadioBtn } from "./RadioBtn";
+
+export { TypesRectangleIconBtn } from "./RectangleIconBtn"
 export type { MetRectangleIconBtnProps } from "./RectangleIconBtn";
 export { default as MetRectangleIconBtn } from "./RectangleIconBtn";
+
+export { TypesRectangleTextBtn, SizeRectangleTextBtn } from "./RectangleTextBtn"
+export type { MetRectangleTextBtnProps } from "./RectangleTextBtn";
+export { default as MetRectangleTextBtn } from "./RectangleTextBtn";
 
 export type { MetSelectProps } from "./Select";
 export { default as MetSelect } from "./Select";
@@ -22,6 +37,9 @@ export { default as MetSquareIconBtn } from "./SquareIconBtn";
 
 export type { MetSquareIconCheckboxProps } from "./SquareIconCheckbox";
 export { default as MetSquareIconCheckbox } from "./SquareIconCheckbox";
+
+export type { MetTabListProps } from "./TabList";
+export { default as MetTabList } from "./TabList";
 
 export type { MetTagButtonProps } from "./Tag/TagButton";
 export { default as MetTagButton } from "./Tag/TagButton";
